requirements.txt
Mastodon.py==1.8.0
PyYAML==6.0
requests==2.22.0
requests-unixsocket==0.2.0
